Extract Entities DTO classes into separate files
