+++ DataModels/Entities/Entities.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Twity.DataModels.Entities
+{
+    [Serializable]
+    public class Entities
+    {
+        public Media[] media;
+        public UserMention[] user_mentions;
+        public HashTag[] hashtags;
+    }
+}
