+++ DataModels/Entities/ExtendedEntities.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Twity.DataModels.Entities
+{
+    [Serializable]
+    public class Extended_Entities
+    {
+        public Media[] media;
+    }
+}
