+++ DataModels/Entities/HashTag.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Twity.DataModels.Entities
+{
+    [Serializable]
+    public class HashTag
+    {
+        public string text;
+    }
+}
