+++ DataModels/Entities/Variant.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Twity.DataModels.Entities
+{
+    [Serializable]
+    public class Variant
+    {
+        public int bitrate;
+        public string content_type;
+        public string url;
+    }
+}
