+++ scripts/twitterJson.cs
@@ -1,68 +1,8 @@
 ﻿using System;
 using Twity.DataModels.Core;
+using Twity.DataModels.Entities;
 
 namespace Twity {
-
-    # region Entities Class
-
-    [Serializable]
-	public class Entities
-    {
-		public Media[] media;
-		public UserMention[] user_mentions;
-		public HashTag[] hashtags;
-	}
-	[Serializable]
-	public class Extended_Entities
-    {
-		public Media[] media;
-	}
-	[Serializable]
-	public class Media
-    {
-		public int id;
-		public int id_str;
-		public string media_url;
-		public string media_url_https;
-		public string type;
-		public Video_Info video_info;
-	}
-	[Serializable]
-	public class UserMention
-    {
-		public long id;
-		public string id_str;
-		public string screen_name;
-		public string name;
-	}
-	[Serializable]
-	public class HashTag
-    {
-		public string text;
-	}
-	[Serializable]
-	public class Symbol
-    {
-		public string text;
-	}
-	[Serializable]
-	public class Video_Info
-    {
-		public long id;
-		public string id_str;
-		public string media_url;
-		public string type;
-		public Variant[] variants;
-	}
-	[Serializable]
-	public class Variant
-    {
-		public int bitrate;
-		public string content_type;
-		public string url;
-	}
-
-    #endregion
 
     #region Rest API Response Class
